refactor : outcome으로 용어 통일 (consume -> outcome)
