--- src/main/java/com/idle/kb_i_dle_backend/consume/mapper/ConsumeMapper.java
@@ -1,7 +1,0 @@
-package com.idle.kb_i_dle_backend.consume.mapper;
-
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface ConsumeMapper {
-}
