+++ src/main/java/com/idle/kb_i_dle_backend/outcome/dto/MonthOutcomeDTO.java
@@ -1,4 +1,4 @@
-package com.idle.kb_i_dle_backend.consume.dto;
+package com.idle.kb_i_dle_backend.outcome.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,7 +9,7 @@ import java.util.List;
 @Getter
 @Setter
 @AllArgsConstructor
-public class MonthConsumeDTO {
+public class MonthOutcomeDTO {
     private int cntMonth;
     private int cntYear;
     List<Long> prices;
