+++ src/main/java/com/idle/kb_i_dle_backend/outcome/dto/OutcomeDTO.java
@@ -1,11 +1,11 @@
-package com.idle.kb_i_dle_backend.consume.dto;
+package com.idle.kb_i_dle_backend.outcome.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class ConsumeDTO {
+public class OutcomeDTO {
     private int wasted;    // 낭비된 금액
     private int monthSum;  // 월 소비 총합
     private int dailyCons; // 일일 소비 금액
